split out some of the non template stuff into cpp file
